refactor: change old-api to functional-api for led api
