+++ src/led/index.ts
@@ -1,2 +1,2 @@
-import * as ledTmp from './old-api';
+import * as ledTmp from './functional-api';
 export const led = ledTmp;
