+++ src/tests/led/run-tests.ts
@@ -1,6 +1,6 @@
 import { executeTest } from '../test-helper';
 // import * as tests from './logi-led';
-import * as oldTests from './old-api';
+import * as functionalTests from './functional-api';
 
 
 export async function runLedTests()
@@ -8,6 +8,6 @@ export async function runLedTests()
 	console.log('Old API');
 	console.log('---');
 
-	await executeTest('getting the sdk version', oldTests.testColor, 10e3);
-	oldTests.shutdown();
+	await executeTest('getting the sdk version', functionalTests.testColor);
+	functionalTests.shutdown();
 }
